aggregator/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Matheus Marabesi <matheus.marabesi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
matplotlib = "^3.7.1"
numpy = "^1.24.2"
pandas = "^1.5.3"
python-slugify = "^8.0.1"

[tool.poetry.dev-dependencies]
pytest-mock = "^3.10.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
